bringing back Step component for easier migration
