+++ packages/stepper/README.md
@@ -27,6 +27,15 @@ npm install @leafygreen-ui/stepper
 </Stepper>
 ```
 
+or
+
+```js
+<Stepper currentStep={0} maxDisplayedSteps={2}>
+  <Step>Step 1</Step>
+  <Step>Step 2</Step>
+</Stepper>
+```
+
 **Output HTML**
 
 ```html
