+++ packages/stepper/src/index.ts
@@ -1,3 +1,5 @@
 import Stepper from './Stepper';
+import Step from './Step';
 export { StepperProps } from './types';
+export { Step };
 export default Stepper;
